bots/gui_messenger/requirements.txt
aiogram
PySide6
qasync
python-dotenv